Implement purchase: migration, model, and controller

+++ app/Http/Controllers/V1/ProductImageController.php
@@ -31,8 +31,10 @@ class ProductImageController extends Controller
     public function destroy($id)
     {
         $image = Image::findOrFail($id);
-        Storage::disk('public')->delete($image->path);
-        $image->delete();
+        if($image->collection_name === 'product_images'){
+            Storage::disk('public')->delete($image->path);
+            $image->delete();
+        }
         
         return ResponseFormatter::success();
     }
